+++ data/src/config.rs
@@ -1,7 +1,9 @@
 use std::fs;
 use std::path::PathBuf;
 
-use rand::Rng;
+use rand::prelude::*;
+use rand_chacha::ChaCha8Rng;
+
 use serde::Deserialize;
 use thiserror::Error;
 
@@ -235,9 +237,7 @@ impl Config {
         }
 
         // Generate a unique nick
-        let mut rng = rand::thread_rng();
-        let rand_digit: u16 = rng.gen_range(1000..=9999);
-        let rand_nick = format!("halloy{rand_digit}");
+        let rand_nick = random_nickname();
 
         // Replace placeholder nick with unique nick
         let config_string = CONFIG_TEMPLATE.replace("__NICKNAME__", rand_nick.as_str());
@@ -248,6 +248,18 @@ impl Config {
 
         let _ = fs::write(config_path, config_bytes);
     }
+}
+
+pub fn random_nickname() -> String {
+    let mut rng = ChaCha8Rng::from_entropy();
+    random_nickname_with_seed(&mut rng)
+}
+
+pub fn random_nickname_with_seed<R: Rng>(rng: &mut R) -> String {
+    let rand_digit: u16 = rng.gen_range(1000..=9999);
+    let rand_nick = format!("halloy{rand_digit}");
+
+    rand_nick
 }
 
 pub fn create_themes_dir() {