+++ ipc/src/lib.rs
@@ -1,0 +1,5 @@
+pub use self::client::connect_and_send;
+pub use self::server::listen;
+
+mod client;
+pub(crate) mod server;