+++ src/main.rs
@@ -11,6 +11,7 @@ mod notification;
 mod screen;
 mod stream;
 mod theme;
+mod url;
 mod widget;
 mod window;
 
@@ -68,7 +69,14 @@ pub fn main() -> iced::Result {
     // before we do any iced related stuff w/ it
     font::set(config_load.as_ref().ok());
 
-    if let Err(error) = Halloy::run(settings(config_load)) {
+    let destination = data::Url::find_in(std::env::args());
+    if let Some(loc) = &destination {
+        if ipc::connect_and_send(loc.to_string()) {
+            return Ok(());
+        }
+    }
+
+    if let Err(error) = Halloy::run(settings(config_load, destination)) {
         log::error!("{}", error.to_string());
         Err(error)
     } else {
@@ -78,7 +86,8 @@ pub fn main() -> iced::Result {
 
 fn settings(
     config_load: Result<Config, config::Error>,
-) -> iced::Settings<Result<Config, config::Error>> {
+    route_received: Option<data::Url>,
+) -> iced::Settings<(Result<Config, config::Error>, Option<data::Url>)> {
     let default_text_size = config_load
         .as_ref()
         .ok()
@@ -93,7 +102,7 @@ fn settings(
             exit_on_close_request: false,
             ..window::settings()
         },
-        flags: config_load,
+        flags: (config_load, route_received),
         id: None,
         antialiasing: false,
         fonts: font::load(),
@@ -192,7 +201,8 @@ pub enum Message {
     Event(Event),
     Tick(Instant),
     Version(Option<String>),
-    CloseModal,
+    Modal(modal::Message),
+    RouteReceived(String),
 }
 
 impl Application for Halloy {
@@ -200,14 +210,20 @@ impl Application for Halloy {
     type Executor = executor::Default;
     type Message = Message;
     type Theme = Theme;
-    type Flags = Result<Config, config::Error>;
+    type Flags = (Result<Config, config::Error>, Option<data::Url>);
 
-    fn new(config_load: Self::Flags) -> (Halloy, Command<Self::Message>) {
-        let (halloy, command) = Halloy::load_from_state(config_load);
+    fn new(flags: Self::Flags) -> (Halloy, Command<Self::Message>) {
+        let (config_load, url_received) = flags;
+
+        let (mut halloy, command) = Halloy::load_from_state(config_load);
         let latest_remote_version =
             Command::perform(version::latest_remote_version(), Message::Version);
 
         let command = Command::batch(vec![command, latest_remote_version]);
+
+        if let Some(url) = url_received {
+            halloy.modal = Some(Modal::RouteReceived(url));
+        }
 
         (halloy, command)
     }
@@ -604,8 +620,46 @@ impl Application for Halloy {
                     Command::none()
                 }
             }
-            Message::CloseModal => {
-                self.modal = None;
+            Message::Modal(message) => {
+                let Some(modal) = &mut self.modal else {
+                    return Command::none();
+                };
+
+                if let Some(event) = modal.update(message) {
+                    match event {
+                        modal::Event::CloseModal => {
+                            self.modal = None;
+                        }
+                        modal::Event::AcceptNewServer => {
+                            if let Some(Modal::RouteReceived(data::Url::ServerConnect {
+                                server,
+                                config,
+                                ..
+                            })) = self.modal.take()
+                            {
+                                let existing_entry = self.servers.entries().find(|entry| {
+                                    entry.server == server || entry.config.server == config.server
+                                });
+
+                                // If server already exists, we only want to join the new channels
+                                if let Some(entry) = existing_entry {
+                                    self.clients.join(&entry.server, &config.channels);
+                                } else {
+                                    self.servers.insert(server, config);
+                                }
+                            }
+                        }
+                    }
+                }
+
+                Command::none()
+            }
+            Message::RouteReceived(route) => {
+                log::info!("RouteRecived: {:?}", route);
+
+                if let Ok(url) = route.parse() {
+                    self.modal = Some(Modal::RouteReceived(url));
+                };
 
                 Command::none()
             }
@@ -627,9 +681,9 @@ impl Application for Halloy {
             .height(Length::Fill)
             .style(theme::container::primary);
 
-        if let Some(modal) = &self.modal {
-            widget::modal(content, modal.view().map(|_| Message::CloseModal), || {
-                Message::CloseModal
+        if let (Some(modal), Screen::Dashboard(_)) = (&self.modal, &self.screen) {
+            widget::modal(content, modal.view().map(Message::Modal), || {
+                Message::Modal(modal::Message::Cancel)
             })
         } else {
             // Align `content` into same view tree shape as `modal`
@@ -656,6 +710,11 @@ impl Application for Halloy {
         )
         .map(Message::Stream);
 
-        Subscription::batch(vec![tick, streams, events().map(Message::Event)])
+        Subscription::batch(vec![
+            url::listen().map(Message::RouteReceived),
+            tick,
+            streams,
+            events().map(Message::Event),
+        ])
     }
 }