+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Added:
 
+- Ability to open `irc://` and `ircs://` URL schemes
 - Ability to overwrite nickname colors by providing a hex string (see [buffer configuration](https://halloy.squidowl.org/configuration/buffer.html#buffernicknamecolor-section)).
 - Ability to overwrite server & internal message colors by providing a hex string (see [buffer configuration](https://halloy.squidowl.org/configuration/buffer.html#bufferserver_messages-section)).